# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shoplike Vintage Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  section: string;
  /** @nullable */
  categoryId?: number | null;
  inStock: boolean;
  /** @nullable */
  stockQty?: number | null;
  /** @nullable */
  sizes?: string[] | null;
  /** @nullable */
  colors?: string[] | null;
  /** @nullable */
  images?: string[] | null;
  /** @nullable */
  videoUrl?: string | null;
  isFeatured: boolean;
  isOnSale: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  section: string;
  /** @nullable */
  categoryId?: number | null;
  inStock: boolean;
  /** @nullable */
  stockQty?: number | null;
  /** @nullable */
  sizes?: string[] | null;
  /** @nullable */
  colors?: string[] | null;
  /** @nullable */
  images?: string[] | null;
  /** @nullable */
  videoUrl?: string | null;
  isFeatured?: boolean;
  isOnSale?: boolean;
}

export interface ProductUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  price?: number;
  /** @nullable */
  comparePrice?: number | null;
  section?: string;
  /** @nullable */
  categoryId?: number | null;
  inStock?: boolean;
  /** @nullable */
  stockQty?: number | null;
  /** @nullable */
  sizes?: string[] | null;
  /** @nullable */
  colors?: string[] | null;
  /** @nullable */
  images?: string[] | null;
  /** @nullable */
  videoUrl?: string | null;
  isFeatured?: boolean;
  isOnSale?: boolean;
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  section: string;
  slug: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryInput {
  name: string;
  section: string;
  slug: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CategoryUpdate {
  name?: string;
  section?: string;
  slug?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface OrderItem {
  productId: number;
  name: string;
  price: number;
  qty: number;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ShippingAddress {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  address: string;
  city: string;
  state: string;
  country: string;
}

export interface Order {
  id: number;
  orderRef: string;
  items: OrderItem[];
  shippingAddress: ShippingAddress;
  total: number;
  status: string;
  paymentStatus: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentRef?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  customerId?: number | null;
  createdAt: string;
}

export interface OrderInput {
  items: OrderItem[];
  shippingAddress: ShippingAddress;
  total: number;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  customerId?: number | null;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderList {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface PaymentInitInput {
  orderId: number;
  amount: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
}

export interface PaymentInitResult {
  checkoutUrl: string;
  reference: string;
}

export interface PaymentVerifyResult {
  status: string;
  amount: number;
  reference: string;
}

export interface CustomerSignupInput {
  name: string;
  email: string;
  password: string;
  /** @nullable */
  phone?: string | null;
}

export interface CustomerLoginInput {
  email: string;
  password: string;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export interface CustomerAuthResult {
  token: string;
  customer: Customer;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminAuthResult {
  token: string;
}

export interface Settings {
  storeName: string;
  /** @nullable */
  storeEmail?: string | null;
  /** @nullable */
  storePhone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  currency: string;
  korapayEnabled: boolean;
  /** @nullable */
  korapayPublicKey?: string | null;
  /** @nullable */
  korapaySecretKey?: string | null;
  /** @nullable */
  korapayEncKey?: string | null;
  /** @nullable */
  smsApiKey?: string | null;
  smsEnabled: boolean;
  /** @nullable */
  emailApiKey?: string | null;
  emailEnabled: boolean;
  /** @nullable */
  bannerText?: string | null;
  bannerEnabled: boolean;
}

export interface SettingsUpdate {
  storeName?: string;
  /** @nullable */
  storeEmail?: string | null;
  /** @nullable */
  storePhone?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  currency?: string;
  korapayEnabled?: boolean;
  /** @nullable */
  korapayPublicKey?: string | null;
  /** @nullable */
  korapaySecretKey?: string | null;
  /** @nullable */
  korapayEncKey?: string | null;
  /** @nullable */
  smsApiKey?: string | null;
  smsEnabled?: boolean;
  /** @nullable */
  emailApiKey?: string | null;
  emailEnabled?: boolean;
  /** @nullable */
  bannerText?: string | null;
  bannerEnabled?: boolean;
}

export interface TopProduct {
  id: number;
  name: string;
  totalSold: number;
  revenue: number;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  pendingOrders: number;
  totalProducts: number;
  todayOrders: number;
  todayRevenue: number;
  topProducts: TopProduct[];
}

export interface SalesDay {
  date: string;
  revenue: number;
  orders: number;
}

export interface ImageFileUpload {
  file: Blob;
}

export interface VideoFileUpload {
  file: Blob;
}

export interface UploadResult {
  url: string;
  filename: string;
}

export type GetProductsParams = {
section?: string;
categoryId?: number;
search?: string;
sale?: boolean;
page?: number;
limit?: number;
};

export type GetFeaturedProductsParams = {
limit?: number;
};

export type DeleteProduct200 = {
  success: boolean;
};

export type GetCategoriesParams = {
section?: string;
};

export type DeleteCategory200 = {
  success: boolean;
};

export type GetOrdersParams = {
status?: string;
page?: number;
limit?: number;
};

export type GetSalesStatsParams = {
days?: number;
};