# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shoplike Vintage Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List categories
      parameters:
        - name: section
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                required: [success]

  # ── Orders ────────────────────────────────────────────────────
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track/{ref}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by reference
      parameters:
        - name: ref
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── Payments (Korapay) ────────────────────────────────────────
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate Korapay payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResult"
  /payments/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify Korapay payment
      parameters:
        - name: reference
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResult"

  # ── Customers ─────────────────────────────────────────────────
  /customers/signup:
    post:
      operationId: customerSignup
      tags: [customers]
      summary: Customer signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerSignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAuthResult"
  /customers/login:
    post:
      operationId: customerLogin
      tags: [customers]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAuthResult"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"

  # ── Settings ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        section: { type: string }
        slug: { type: string }
        imageUrl: { type: ["string", "null"] }
      required: [id, name, section, slug]
    CategoryInput:
      type: object
      properties:
        name: { type: string }
        section: { type: string }
        slug: { type: string }
        imageUrl: { type: ["string", "null"] }
      required: [name, section, slug]
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        section: { type: string }
        slug: { type: string }
        imageUrl: { type: ["string", "null"] }

    # ── Order ─────────────────────────────────────────────────────
    OrderItem:
      type: object
      properties:
        productId: { type: integer }
        name: { type: string }
        price: { type: number }
        qty: { type: integer }
        size: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
      required: [productId, name, price, qty]
    ShippingAddress:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        country: { type: string }
      required: [fullName, phone, address, city, state, country]
    Order:
      type: object
      properties:
        id: { type: integer }
        orderRef: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        total: { type: number }
        status: { type: string }
        paymentStatus: { type: string }
        paymentMethod: { type: ["string", "null"] }
        paymentRef: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        customerId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, orderRef, items, shippingAddress, total, status, paymentStatus, createdAt]
    OrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        total: { type: number }
        notes: { type: ["string", "null"] }
        paymentMethod: { type: ["string", "null"] }
        customerId: { type: ["integer", "null"] }
      required: [items, shippingAddress, total]
    OrderStatusUpdate:
      type: object
      properties:
        status: { type: string }
      required: [status]
    OrderList:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [orders, total, page, limit]

    # ── Payment ──────────────────────────────────────────────────
    PaymentInitInput:
      type: object
      properties:
        orderId: { type: integer }
        amount: { type: number }
        email: { type: string }
        name: { type: string }
        phone: { type: ["string", "null"] }
      required: [orderId, amount, email, name]
    PaymentInitResult:
      type: object
      properties:
        checkoutUrl: { type: string }
        reference: { type: string }
      required: [checkoutUrl, reference]
    PaymentVerifyResult:
      type: object
      properties:
        status: { type: string }
        amount: { type: number }
        reference: { type: string }
      required: [status, amount, reference]

    # ── Customer ─────────────────────────────────────────────────
    CustomerSignupInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        phone: { type: ["string", "null"] }
      required: [name, email, password]
    CustomerLoginInput:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    Customer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, name, email, createdAt]
    CustomerAuthResult:
      type: object
      properties:
        token: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
      required: [token, customer]

    # ── Admin ─────────────────────────────────────────────────────
    AdminLoginInput:
      type: object
      properties:
        password: { type: string }
      required: [password]
    AdminAuthResult:
      type: object
      properties:
        token: { type: string }
      required: [token]

    # ── Settings ─────────────────────────────────────────────────
...