# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shoplike Vintage Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { ShippingAddress } from './shippingAddress';

export interface Order {
  id: number;
  orderRef: string;
  items: OrderItem[];
  shippingAddress: ShippingAddress;
  total: number;
  status: string;
  paymentStatus: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentRef?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  customerId?: number | null;
  createdAt: string;
}