# lib/api-zod/src/generated/types/paymentInitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shoplike Vintage Store API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInitResult {
  checkoutUrl: string;
  reference: string;
}