# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shoplike Vintage Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingAddress {
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  address: string;
  city: string;
  state: string;
  country: string;
}